+++ packages/building-rollup/demo/js/rollup.modern.config.js
@@ -1,20 +1,5 @@
-const cpy = require('rollup-plugin-cpy');
 const createDefaultConfig = require('../../modern-config');
 
-const config = createDefaultConfig({
+module.exports = createDefaultConfig({
   input: './demo/js/index.html',
 });
-
-module.exports = {
-  ...config,
-  plugins: [
-    ...config.plugins,
-    cpy({
-      files: ['**/*.txt'],
-      dest: 'dist',
-      options: {
-        parents: true,
-      },
-    }),
-  ],
-};