+++ packages/create/test/integration.test.js
@@ -6,6 +6,7 @@ import { promisify } from 'util';
 import { lstatSync, readdirSync, readFileSync } from 'fs';
 import { join } from 'path';
 import { CLIEngine } from 'eslint';
+import { generateCommand } from './generate-command.js';
 
 const exec = promisify(_exec);
 
@@ -16,8 +17,6 @@ const { expect } = chai;
 chai.use(chaiFs);
 
 const getFileMessages = ({ messages, filePath }) => `${filePath}:\n${messages.join('\n')}`;
-
-const COMMAND_PATH = join(__dirname, '../src/create.js');
 
 const ACTUAL_PATH = join(process.cwd(), './scaffold-app');
 
@@ -81,18 +80,7 @@ describe('create', () => {
 
   const expectedPath = join(__dirname, './snapshots/fully-loaded-app');
 
-  const command = `node -r @babel/register \
-    ${COMMAND_PATH} \
-      --destinationPath ${destinationPath} \
-      --type scaffold \
-      --scaffoldType app \
-      --features linting testing demoing building \
-      --buildingType rollup \
-      --scaffoldFilesFor testing demoing building \
-      --tagName scaffold-app \
-      --writeToDisk true \
-      --installDependencies false
-  `;
+  const command = generateCommand({ destinationPath });
 
   before(generate({ command, expectedPath }));