+++ packages/es-dev-server/demo/compatibility/module-features.js
@@ -1,5 +1,0 @@
-import './module-features-a.js';
-
-import('./module-features-b.js');
-
-console.log('import.meta.url', import.meta.url);