+++ packages/rollup-plugin-index-html/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.2](https://github.com/open-wc/open-wc/compare/rollup-plugin-index-html@1.7.1...rollup-plugin-index-html@1.7.2) (2019-11-24)
+
+**Note:** Version bump only for package rollup-plugin-index-html
+
+
+
+
+
 ## [1.7.1](https://github.com/open-wc/open-wc/compare/rollup-plugin-index-html@1.7.0...rollup-plugin-index-html@1.7.1) (2019-11-20)
 
 **Note:** Version bump only for package rollup-plugin-index-html